requirements.txt
# Updated requirements.txt
prometheus-client==0.20.0
faker==24.11.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
statsmodels==0.14.0
tensorflow==2.15.0
prophet==1.1.5
xgboost==2.0.3
sentence-transformers==2.6.1
faiss-cpu==1.8.0
streamlit==1.32.0
plotly==5.19.0
ollama==0.1.6